Feat: Domain 이 Application 에 의존하는 경우 수정
